--- src/main/java/tasktwo/Repository.java
@@ -1,9 +1,0 @@
-package tasktwo;
-
-/**
- * @author Petar Nedelchev <peter.krasimirov@gmail.com>
- */
-public interface Repository {
-    void createRepository();
-    void deleteRepository();
-}
